requirements.txt
notebook==6.4.12
boto3==1.24.75
requests==2.28.1
python-dotenv==0.21.0
retry==0.9.2
dune-client
pandas==1.4.4
streamlit==1.12.2
matplotlib==3.6.0
tqdm==4.64.1
prophet==1.1.1
spacy==3.4.1
spacytextblob==4.0.0